refacto(app): move ui internal code
