+++ packages/~/users/api/tsconfig.json
@@ -6,6 +6,7 @@
   "references": [
     { "path": "../../app/layout/tsconfig.json" },
     { "path": "../../app/middleware/tsconfig.json" },
+    { "path": "../../app/ui/tsconfig.json" },
     { "path": "../../app/urls/tsconfig.json" },
     { "path": "../../moderations/repository/tsconfig.json" },
     { "path": "../../organizations/repository/tsconfig.json" },
