--- src/ui/loader/Loader.tsx
@@ -1,20 +1,0 @@
-//
-
-import type { VariantProps } from "tailwind-variants";
-import { loader } from ".";
-
-export function Loader(
-  props: Hono.HTMLAttributes & VariantProps<typeof loader>,
-) {
-  const { class: className, htmx_indicator, ...other_props } = props;
-  const { base, span } = loader({ htmx_indicator });
-  return (
-    <div
-      class={base({ className: String(className) })}
-      role="status"
-      {...other_props}
-    >
-      <span class={span()}>Loading...</span>
-    </div>
-  );
-}
